Add support for arrows and fields in comments
Chessbase software is generating comments in a way that is not supported
by this parser.

`{ [%cal Rc4c6] comment about move }`
`{ [%csl Yd4] comment about move }`

With updated rules comments after arrows and square colouring will be
added to `commentAfter`. I think it is the best place and it will not
break HTML generator for the UI.